annotations now work, the result is just printed right now
